requirements.txt
httplib2==0.10.3
python-telegram-bot==9.0.0
pyzomato==0.5
requests==2.11.1
urllib3==1.22

